+++ CHANGELOG.md
@@ -4,4 +4,8 @@
 
 ## [Unreleased]
 
-Test suite: 656 -> 656
+Test suite: 656 -> 674
+
+### Added
+
+- Viewer: multi-select filter on the **All analyses** table.